add custom hook for handling image source url
